Upgrade datadaog-metrics, rename decorators -> tracing
