+++ package.json
@@ -85,7 +85,7 @@
     "copy-to-clipboard": "^3.3.3",
     "core-js": "^3.26.1",
     "crypto-js": "^4.1.1",
-    "datadog-metrics": "^0.9.3",
+    "datadog-metrics": "^0.10.2",
     "date-fns": "^2.25.0",
     "dd-trace": "^3.9.3",
     "dotenv": "^4.0.0",
