+++ server/collaboration/AuthenticationExtension.ts
@@ -1,5 +1,5 @@
 import { onAuthenticatePayload, Extension } from "@hocuspocus/server";
-import { trace } from "@server/logging/decorators";
+import { trace } from "@server/logging/tracing";
 import Document from "@server/models/Document";
 import { can } from "@server/policies";
 import { getUserForJWT } from "@server/utils/jwt";
