+++ server/collaboration/PersistenceExtension.ts
@@ -7,7 +7,7 @@ import {
 import * as Y from "yjs";
 import { sequelize } from "@server/database/sequelize";
 import Logger from "@server/logging/Logger";
-import { trace } from "@server/logging/decorators";
+import { trace } from "@server/logging/tracing";
 import Document from "@server/models/Document";
 import documentCollaborativeUpdater from "../commands/documentCollaborativeUpdater";
 import markdownToYDoc from "./utils/markdownToYDoc";
