+++ server/logging/Logger.ts
@@ -5,7 +5,7 @@ import winston from "winston";
 import env from "@server/env";
 import Metrics from "@server/logging/Metrics";
 import Sentry from "@server/logging/sentry";
-import * as Tracing from "./tracing";
+import * as Tracing from "./tracer";
 
 const isProduction = env.ENVIRONMENT === "production";
