+++ server/presenters/document.ts
@@ -1,4 +1,4 @@
-import { traceFunction } from "@server/logging/decorators";
+import { traceFunction } from "@server/logging/tracing";
 import { Document } from "@server/models";
 import DocumentHelper from "@server/models/helpers/DocumentHelper";
 import presentUser from "./user";
