+++ server/services/websockets.ts
@@ -6,8 +6,8 @@ import IO from "socket.io";
 import { createAdapter } from "socket.io-redis";
 import Logger from "@server/logging/Logger";
 import Metrics from "@server/logging/Metrics";
-import { traceFunction } from "@server/logging/decorators";
-import * as Tracing from "@server/logging/tracing";
+import * as Tracing from "@server/logging/tracer";
+import { traceFunction } from "@server/logging/tracing";
 import { Document, Collection, View, User } from "@server/models";
 import { can } from "@server/policies";
 import { getUserForJWT } from "@server/utils/jwt";
